Edited loading another profile

+++ coursework/app/src/main/java/ru/fasdev/tfs/screen/fragment/people/PeopleFragment.kt
@@ -109,7 +109,7 @@ class PeopleFragment :
     }
 
     override fun onClickUser(idUser: Int, email: String) {
-        rootRouter.navigateTo(ProfileAnotherFragment.getScreen(idUser))
+        rootRouter.navigateTo(ProfileAnotherFragment.getScreen(idUser.toLong()))
         binding.searchLayout.hideKeyboard()
     }
